Edit mode changes
Added light dividers between scenes so that these breaks can be more easily seen - especially when pasting text in and you're not sure where the breaks are occurring.
Fixed bug when selecting all text in a scene, and trying to deleting it, the text ended up being copied to the previous scene instead of being actually deleted.